python/requirements.txt
fastapi
uvicorn
pymysql
pydantic
python-dotenv
pandas
numpy
scikit-learn
tensorflow==2.20.0
